requirements.txt
openai>=1.0
anthropic>=0.40
google-genai>=0.3
python-dotenv>=1.0
tabulate>=0.9
matplotlib>=3.8
numpy>=1.24
pytest>=8.0
